+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `StatdDb` and `Map::new`. They are only used internally and shouldn't be
   called from another crate.
 - `Map::update_account`. Use `Table<Account>::update_account` instead.
+- `BlockingPgPool` has been removed as `diesel_async` is used instead.
 
 ## [0.4.0] - 2023-04-06